decoupled accepted devices tab & list
Changelog: None
